Fix proposal for the official image
It is mostly codestyle. (I really like it like this :) )

I moved `gpg` one line upper as it will change less than the `RC_VERSION`.

I added the env var `PORT` if not, meteor was not starting.

I changed the port to 3000 as node will not have the right to start on port 80 from non root.
This part is a bit annoying, as it will not work 'out of the box'.
I tried to find a solution, but there is nothing satisfactory. The best might be to start an nginx in front in the `docker-compose.yml`.
In the mean time on my infrastructure, I just run it as root (like most people do on the Internet apparently).

This is sad.

I also looked at [this](http://stackoverflow.com/a/27805105/1703029) without success :/

I'll ask on StackOverflow, I'll report here if I get responses!

All the best